Added title bar to the motd window.
I've gotten multiple requests from people to add a title bar to
the motd window so they don't have to scroll down to close it.
I agree, I like to be able to just click the X as well, but if
Head wanted it without the title bar for a reason, feel free to
change it back!