My turn at trying to fix flickering.  Should be working now.
However I did change how it works.  I ran this by Abi the other day
but I'm not sure I was really clear in what I meant.  Basically
It's no longer it's own proc or verb, it's now part of the Click()
proc for lights, so ghosts simply have to be next to the light and
just click on it and it flickers.  Much easier/better for everyone
in my opinion.  It didn't seem to cause any problems with the quick
testing I did, but if you guys hate it or foresee issues arising
in the future due to it, feel free to change it.