Hidden tabs tracked, tab recovery on connect, OOC initially visible
On client/New, it will search for existing tabs and add them as hidden or shown.
This covers the case where a tab is added directly to the skin, and when a client is still connected after a reboot.
As a side effect of testing, and intentionally left in, OOC will be added in this manner. I witnessed a small bit of confusion once where someone was wondering where it had went until they commented on it in OOC, causing it to appear. This way, nobody has to complain about OOC, since it is probably the second most important tab to the average user.

It cannot recover channel preferences unless it stores those directly between rounds (not a bad idea)

Also included is five potential asteroid icons, though I will probably rework them somewhat, so don't get too attached to the corners.