hiding accessories under suits
Essentially a whole new system for hiding accessories under suits.
Every suit now has assigned flags that determine which bodypart it covers (pretty much based on in-game sprites and common sense). Every accessory has assigned flags that determine which bodypart it belongs to. If an accessory belongs to a bodypart that is covered by the suit the user is wearing, that accessory does not show up in examine.